move nanotime to commons
