+++ atlasdb-commons/src/main/java/com/palantir/common/time/NanoTime.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.leader.lease;
+package com.palantir.common.time;
 
 import java.time.Duration;
 import java.util.Objects;
@@ -25,8 +25,7 @@ import com.google.common.annotations.VisibleForTesting;
 public final class NanoTime implements Comparable<NanoTime> {
     private final long time;
 
-    @VisibleForTesting
-    NanoTime(long time) {
+    public NanoTime(long time) {
         this.time = time;
     }
